requirements.txt
celery
Django
djangorestframework
djangorestframework-jwt
Pillow
PyJWT
redis
